PRLC100 Rebuild Added Callbacks for Citizen Case Submission
